Modified AbsModel
